put queue safety checks into new file
